+++ BaseLib/OperatorDecrement_Functions.inl
@@ -6,5 +6,5 @@ void DecrementDouble(void* pvDest) { (*(double*)pvDest)--; }
 void DecrementUInt(void* pvDest) { (*(unsigned int*)pvDest)--; }
 void DecrementUShort(void* pvDest) { (*(uint16*)pvDest)--; }
 void DecrementUChar(void* pvDest) { (*(unsigned char*)pvDest)--; }
-void DecrementLong(void* pvDest) { (*(long long int*)pvDest)--; }
-void DecrementULong(void* pvDest) { (*(unsigned long long int*)pvDest)--; }
+void DecrementLong(void* pvDest) { (*(int64*)pvDest)--; }
+void DecrementULong(void* pvDest) { (*(uint64*)pvDest)--; }
