+++ BaseLib/OperatorEqualNumber_Functions.inl
@@ -6,8 +6,8 @@ void EqualIntDouble(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool
 void EqualIntUInt(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(int*)pvLeft == (int)(*(unsigned int*)pvRight)); }
 void EqualIntUShort(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(int*)pvLeft == *(uint16*)pvRight); }
 void EqualIntUChar(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(int*)pvLeft == *(unsigned char*)pvRight); }
-void EqualIntLong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(int*)pvLeft == *(long long int*)pvRight); }
-void EqualIntULong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(int*)pvLeft == *(unsigned long long int*)pvRight); }
+void EqualIntLong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(int*)pvLeft == *(int64*)pvRight); }
+void EqualIntULong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(int*)pvLeft == *(uint64*)pvRight); }
 void EqualShortInt(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(int16*)pvLeft == *(int*)pvRight); }
 void EqualShortShort(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(int16*)pvLeft == *(int16*)pvRight); }
 void EqualShortChar(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(int16*)pvLeft == *(char*)pvRight); }
@@ -16,8 +16,8 @@ void EqualShortDouble(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bo
 void EqualShortUInt(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(int16*)pvLeft == (int16)(*(unsigned int*)pvRight)); }
 void EqualShortUShort(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(int16*)pvLeft == *(uint16*)pvRight); }
 void EqualShortUChar(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(int16*)pvLeft == *(unsigned char*)pvRight); }
-void EqualShortLong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(int16*)pvLeft == *(long long int*)pvRight); }
-void EqualShortULong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(int16*)pvLeft == *(unsigned long long int*)pvRight); }
+void EqualShortLong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(int16*)pvLeft == *(int64*)pvRight); }
+void EqualShortULong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(int16*)pvLeft == *(uint64*)pvRight); }
 void EqualCharInt(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(char*)pvLeft == *(int*)pvRight); }
 void EqualCharShort(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(char*)pvLeft == *(int16*)pvRight); }
 void EqualCharChar(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(char*)pvLeft == *(char*)pvRight); }
@@ -26,8 +26,8 @@ void EqualCharDouble(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (boo
 void EqualCharUInt(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(char*)pvLeft == (char)(*(unsigned int*)pvRight)); }
 void EqualCharUShort(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(char*)pvLeft == *(uint16*)pvRight); }
 void EqualCharUChar(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(char*)pvLeft == *(unsigned char*)pvRight); }
-void EqualCharLong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(char*)pvLeft == *(long long int*)pvRight); }
-void EqualCharULong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(char*)pvLeft == *(unsigned long long int*)pvRight); }
+void EqualCharLong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(char*)pvLeft == *(int64*)pvRight); }
+void EqualCharULong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(char*)pvLeft == *(uint64*)pvRight); }
 void EqualFloatInt(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(float*)pvLeft == *(int*)pvRight); }
 void EqualFloatShort(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(float*)pvLeft == *(int16*)pvRight); }
 void EqualFloatChar(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(float*)pvLeft == *(char*)pvRight); }
@@ -36,8 +36,8 @@ void EqualFloatDouble(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bo
 void EqualFloatUInt(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(float*)pvLeft == *(unsigned int*)pvRight); }
 void EqualFloatUShort(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(float*)pvLeft == *(uint16*)pvRight); }
 void EqualFloatUChar(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(float*)pvLeft == *(unsigned char*)pvRight); }
-void EqualFloatLong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(float*)pvLeft == *(long long int*)pvRight); }
-void EqualFloatULong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(float*)pvLeft == *(unsigned long long int*)pvRight); }
+void EqualFloatLong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(float*)pvLeft == *(int64*)pvRight); }
+void EqualFloatULong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(float*)pvLeft == *(uint64*)pvRight); }
 void EqualDoubleInt(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(double*)pvLeft == *(int*)pvRight); }
 void EqualDoubleShort(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(double*)pvLeft == *(int16*)pvRight); }
 void EqualDoubleChar(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(double*)pvLeft == *(char*)pvRight); }
@@ -46,8 +46,8 @@ void EqualDoubleDouble(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (b
 void EqualDoubleUInt(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(double*)pvLeft == *(unsigned int*)pvRight); }
 void EqualDoubleUShort(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(double*)pvLeft == *(uint16*)pvRight); }
 void EqualDoubleUChar(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(double*)pvLeft == *(unsigned char*)pvRight); }
-void EqualDoubleLong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(double*)pvLeft == *(long long int*)pvRight); }
-void EqualDoubleULong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(double*)pvLeft == *(unsigned long long int*)pvRight); }
+void EqualDoubleLong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(double*)pvLeft == *(int64*)pvRight); }
+void EqualDoubleULong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(double*)pvLeft == *(uint64*)pvRight); }
 void EqualUIntInt(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(unsigned int*)pvLeft == (unsigned int)(*(int*)pvRight)); }
 void EqualUIntShort(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(unsigned int*)pvLeft == (unsigned int)(*(int16*)pvRight)); }
 void EqualUIntChar(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(unsigned int*)pvLeft == (unsigned int)(*(char*)pvRight)); }
@@ -56,8 +56,8 @@ void EqualUIntDouble(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (boo
 void EqualUIntUInt(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(unsigned int*)pvLeft == *(unsigned int*)pvRight); }
 void EqualUIntUShort(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(unsigned int*)pvLeft == *(uint16*)pvRight); }
 void EqualUIntUChar(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(unsigned int*)pvLeft == *(unsigned char*)pvRight); }
-void EqualUIntLong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(unsigned int*)pvLeft == *(long long int*)pvRight); }
-void EqualUIntULong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(unsigned int*)pvLeft == *(unsigned long long int*)pvRight); }
+void EqualUIntLong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(unsigned int*)pvLeft == *(int64*)pvRight); }
+void EqualUIntULong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(unsigned int*)pvLeft == *(uint64*)pvRight); }
 void EqualUShortInt(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(uint16*)pvLeft == *(int*)pvRight); }
 void EqualUShortShort(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(uint16*)pvLeft == *(int16*)pvRight); }
 void EqualUShortChar(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(uint16*)pvLeft == *(char*)pvRight); }
@@ -66,8 +66,8 @@ void EqualUShortDouble(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (b
 void EqualUShortUInt(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(uint16*)pvLeft == *(unsigned int*)pvRight); }
 void EqualUShortUShort(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(uint16*)pvLeft == *(uint16*)pvRight); }
 void EqualUShortUChar(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(uint16*)pvLeft == *(unsigned char*)pvRight); }
-void EqualUShortLong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(uint16*)pvLeft == *(long long int*)pvRight); }
-void EqualUShortULong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(uint16*)pvLeft == *(unsigned long long int*)pvRight); }
+void EqualUShortLong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(uint16*)pvLeft == *(int64*)pvRight); }
+void EqualUShortULong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(uint16*)pvLeft == *(uint64*)pvRight); }
 void EqualUCharInt(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(unsigned char*)pvLeft == *(int*)pvRight); }
 void EqualUCharShort(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(unsigned char*)pvLeft == *(int16*)pvRight); }
 void EqualUCharChar(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(unsigned char*)pvLeft == *(char*)pvRight); }
@@ -76,25 +76,25 @@ void EqualUCharDouble(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bo
 void EqualUCharUInt(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(unsigned char*)pvLeft == *(unsigned int*)pvRight); }
 void EqualUCharUShort(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(unsigned char*)pvLeft == *(uint16*)pvRight); }
 void EqualUCharUChar(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(unsigned char*)pvLeft == *(unsigned char*)pvRight); }
-void EqualUCharLong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(unsigned char*)pvLeft == *(long long int*)pvRight); }
-void EqualUCharULong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(unsigned char*)pvLeft == *(unsigned long long int*)pvRight); }
-void EqualLongInt(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(long long int*)pvLeft == *(int*)pvRight); }
-void EqualLongShort(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(long long int*)pvLeft == *(int16*)pvRight); }
-void EqualLongChar(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(long long int*)pvLeft == *(char*)pvRight); }
-void EqualLongFloat(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(long long int*)pvLeft == *(float*)pvRight); }
-void EqualLongDouble(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(long long int*)pvLeft == *(double*)pvRight); }
-void EqualLongUInt(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(long long int*)pvLeft == *(unsigned int*)pvRight); }
-void EqualLongUShort(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(long long int*)pvLeft == *(uint16*)pvRight); }
-void EqualLongUChar(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(long long int*)pvLeft == *(unsigned char*)pvRight); }
-void EqualLongLong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(long long int*)pvLeft == *(long long int*)pvRight); }
-void EqualLongULong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(long long int*)pvLeft == (long long int)(*(unsigned long long int*)pvRight)); }
-void EqualULongInt(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(unsigned long long int*)pvLeft == *(int*)pvRight); }
-void EqualULongShort(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(unsigned long long int*)pvLeft == *(int16*)pvRight); }
-void EqualULongChar(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(unsigned long long int*)pvLeft == *(char*)pvRight); }
-void EqualULongFloat(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(unsigned long long int*)pvLeft == *(float*)pvRight); }
-void EqualULongDouble(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(unsigned long long int*)pvLeft == *(double*)pvRight); }
-void EqualULongUInt(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(unsigned long long int*)pvLeft == *(unsigned int*)pvRight); }
-void EqualULongUShort(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(unsigned long long int*)pvLeft == *(uint16*)pvRight); }
-void EqualULongUChar(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(unsigned long long int*)pvLeft == *(unsigned char*)pvRight); }
-void EqualULongLong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(unsigned long long int*)pvLeft == (unsigned long long int)(*(long long int*)pvRight)); }
-void EqualULongULong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(unsigned long long int*)pvLeft == *(unsigned long long int*)pvRight); }
+void EqualUCharLong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(unsigned char*)pvLeft == *(int64*)pvRight); }
+void EqualUCharULong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(unsigned char*)pvLeft == *(uint64*)pvRight); }
+void EqualLongInt(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(int64*)pvLeft == *(int*)pvRight); }
+void EqualLongShort(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(int64*)pvLeft == *(int16*)pvRight); }
+void EqualLongChar(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(int64*)pvLeft == *(char*)pvRight); }
+void EqualLongFloat(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(int64*)pvLeft == *(float*)pvRight); }
+void EqualLongDouble(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(int64*)pvLeft == *(double*)pvRight); }
+void EqualLongUInt(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(int64*)pvLeft == *(unsigned int*)pvRight); }
+void EqualLongUShort(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(int64*)pvLeft == *(uint16*)pvRight); }
+void EqualLongUChar(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(int64*)pvLeft == *(unsigned char*)pvRight); }
+void EqualLongLong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(int64*)pvLeft == *(int64*)pvRight); }
+void EqualLongULong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(int64*)pvLeft == (int64)(*(uint64*)pvRight)); }
+void EqualULongInt(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(uint64*)pvLeft == *(int*)pvRight); }
+void EqualULongShort(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(uint64*)pvLeft == *(int16*)pvRight); }
+void EqualULongChar(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(uint64*)pvLeft == *(char*)pvRight); }
+void EqualULongFloat(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(uint64*)pvLeft == *(float*)pvRight); }
+void EqualULongDouble(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(uint64*)pvLeft == *(double*)pvRight); }
+void EqualULongUInt(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(uint64*)pvLeft == *(unsigned int*)pvRight); }
+void EqualULongUShort(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(uint64*)pvLeft == *(uint16*)pvRight); }
+void EqualULongUChar(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(uint64*)pvLeft == *(unsigned char*)pvRight); }
+void EqualULongLong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(uint64*)pvLeft == (uint64)(*(int64*)pvRight)); }
+void EqualULongULong(bool* pvDest, void* pvLeft, void* pvRight) { *pvDest = (bool)(*(uint64*)pvLeft == *(uint64*)pvRight); }
